[DON'T MERGE] Avoid accessing the non-existent availableDates webpage
Problem:
As of today 2023/04/22, the following code fails to access the desired webpage (which contains a list of available dates of openings) due to the webpage doesn't exist:
```node
const response = await targetPage.goto(
  'https://ais.usvisa-info.com/en-' + region + '/niv/schedule/' +
  appointmentId + '/appointment/days/' + consularId +
  '.json?appointments[expedite]=false'
);
```

Solution:
I don't know what is the alternative webpage. Instead, I bypassed it by a workaround -- Count the number of clicks on the calendar's next button. The number of clicks is approximately the number of months to the next opening.

BIG WARNING1:
This approach has a 1-month error --- If you enter the expected date of (m=X,d=Y),
the program selects extra days between (m=X,d=Y) to (m=X,d=31).